requirements.txt
asgiref==3.8.1
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.7
Django==5.0.6
django-fernet-fields-v2==0.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
